first pipe link using grouping and fragment ideia

+++ timberborn-modding/Assets/Mods/Pipe/Scripts/PipeWaterInput.cs
@@ -3,6 +3,10 @@ using Timberborn.BaseComponentSystem;
 using Timberborn.WaterSystem;
 using Timberborn.BlockSystem;
 using UnityEngine;
+using Timberborn.CoreUI;
+using Timberborn.MechanicalSystem;
+using System.Linq;
+using Timberborn.SingletonSystem;
 
 namespace Mods.Pipe.Scripts
 {
@@ -10,12 +14,18 @@ namespace Mods.Pipe.Scripts
   {
 
     [SerializeField]
-    private string _inputCoordinates; // Front Back , Left Right , Top Botton
+    private string _positionLabel; // Front Back , Left Right , Top Botton
 
     [SerializeField]
     private Vector3Int _waterCoordinates;
 
+    public int direction { get; private set; } = 0; // 0 = stop , 1 = enter , 2 = exit
+
+    private PipeWaterLink _pipeWaterLink;
+
     private BlockObject _blockObject;
+
+    private BlockService _blockService;
 
     private Vector3Int _waterCoordinatesTransformed;
 
@@ -23,43 +33,70 @@ namespace Mods.Pipe.Scripts
 
     private IThreadSafeWaterMap _threadSafeWaterMap;
 
+    public float Water { get; private set; }
+
+    public float CleanWater { get; private set; }
+
+    public float ContaminatedWater { get; private set; }
+
+    public bool inputEnabled { get; private set; } = false;
+
     public bool IsUnderwater
     {
       get
       {
+        if (_threadSafeWaterMap == null)
+          return false;
         return _threadSafeWaterMap.CellIsUnderwater(_waterCoordinatesTransformed);
       }
     }
 
-    public float AvailableWater
+    private float _AvailableWater
     {
       get
       {
+        if (_threadSafeWaterMap == null)
+          return 0f;
         float totalWaterAmount = _threadSafeWaterMap.WaterHeightOrFloor(_waterCoordinatesTransformed);
         return totalWaterAmount - (float)(_waterCoordinatesTransformed.z);
       }
     }
 
-    public float AvailableContaminatedWater
+    private float _AvailableContaminatedWater
     {
       get
       {
-        float availableWater = this.AvailableWater;
+        if (_threadSafeWaterMap == null)
+          return 0f;
+        float availableWater = this.Water;
         float contaminationPercentage = _threadSafeWaterMap.ColumnContamination(_waterCoordinatesTransformed);
         return availableWater * contaminationPercentage;
       }
     }
 
+    public void UpdateAvailableWaters()
+    {
+      Water = _AvailableWater;
+      ContaminatedWater = _AvailableContaminatedWater;
+      CleanWater = Water - ContaminatedWater;
+    }
+
     [Inject]
-    public void InjectDependencies(IWaterService waterService, IThreadSafeWaterMap threadSafeWaterMap)
+    public void InjectDependencies(
+      IWaterService waterService,
+      IThreadSafeWaterMap threadSafeWaterMap,
+      BlockService blockService
+    )
     {
       _waterService = waterService;
       _threadSafeWaterMap = threadSafeWaterMap;
+      _blockService = blockService;
     }
 
     public void Awake()
     {
       _blockObject = GetComponentFast<BlockObject>();
+      _pipeWaterLink = GetComponentFast<PipeWaterLink>();
     }
 
     public void Start()
@@ -67,8 +104,58 @@ namespace Mods.Pipe.Scripts
       _waterCoordinatesTransformed = _blockObject.Transform(_waterCoordinates);
     }
 
+    public void CheckClearInput()
+    {
+      var Pipelink = _blockService.GetObjectsWithComponentAt<PipeWaterLink>(Vector3Int.FloorToInt(_waterCoordinatesTransformed)).ToList().First();
+      if (Pipelink)
+      {
+        inputEnabled = false;
+        Pipelink.Connect(_pipeWaterLink);
+        inputEnabled = true;
+        return;
+      }
+      var block = _blockService.GetObjectsWithComponentAt<BlockObject>(Vector3Int.FloorToInt(_waterCoordinatesTransformed)).ToList().First();
+      inputEnabled = !block.Solid;
+      inputEnabled = true;
+    }
+
+    public void StopWater()
+    {
+      direction = 0;
+    }
+
+    public void InWater()
+    {
+      direction = 1;
+    }
+
+    public void OutWater()
+    {
+      direction = 2;
+    }
+
+    public void AddWater(float cleanWater, float contaminatedWater)
+    {
+      if (!inputEnabled)
+      {
+        return;
+      }
+      if (cleanWater > 0f)
+      {
+        _waterService.AddCleanWater(_waterCoordinatesTransformed, cleanWater);
+      }
+      if (contaminatedWater > 0f)
+      {
+        _waterService.AddContaminatedWater(_waterCoordinatesTransformed, contaminatedWater);
+      }
+    }
+
     public void RemoveWater(float cleanWater, float contaminatedWater)
     {
+      if (!inputEnabled)
+      {
+        return;
+      }
       if (cleanWater > 0f)
       {
         _waterService.RemoveCleanWater(_waterCoordinatesTransformed, cleanWater);
@@ -79,16 +166,29 @@ namespace Mods.Pipe.Scripts
       }
     }
 
-    public void AddWater(float cleanWater, float contaminatedWater)
+    public void MoveWater(float cleanWater, float contaminatedWater)
     {
-      if (cleanWater > 0f)
+      bool isAddWater = cleanWater >= 0f && contaminatedWater >= 0f;
+      cleanWater = Mathf.Abs(cleanWater);
+      contaminatedWater = Mathf.Abs(contaminatedWater);
+      if (isAddWater)
       {
-        _waterService.AddCleanWater(_waterCoordinatesTransformed, cleanWater);
+        AddWater(cleanWater, contaminatedWater);
       }
-      if (contaminatedWater > 0f)
+      else
       {
-        _waterService.AddContaminatedWater(_waterCoordinatesTransformed, contaminatedWater);
+        RemoveWater(cleanWater, contaminatedWater);
       }
+    }
+
+    public string GetInfo()
+    {
+      string info = "{ ";
+      info += "label = " + _positionLabel;
+      info += "enabled = " + inputEnabled;
+      info += "water = " + Water.ToString("0.00");
+      info += " }";
+      return info;
     }
   }
 }
